feat: move UtxoItem to another component
