+++ packages/app-explorer/src/systems/Core/components/UtxoItem/constants.ts
@@ -1,0 +1,2 @@
+export const UTXO_ID_SIZE = 34;
+export const TX_ID_SIZE = 32;
