+++ packages/app-explorer/src/systems/Core/components/UtxoItem/types.ts
@@ -1,0 +1,9 @@
+import type { CSSProperties } from 'react';
+import type { UtxoItemType } from '~/systems/Core/components/Utxos/types';
+
+export type UtxoItemProps = {
+  item: UtxoItemType;
+  assetId?: string;
+  style?: CSSProperties;
+  index: number;
+};
